\name{rs.clearCache}
\alias{rs.clearCache}

\title{Clears Fingerprint Cache}

\description{
Clears fingerprint cache. \code{rs.clearCache} should be called before fingerprint
method (or related properties) are modified.
}

\usage{
rs.clearCache()
}

\value{
  No return value
}

\author{
Varun Giri \email{varungiri@gmail.com}
}

\seealso{
\code{\link[RxnSim:rs.compute]{rs.compute}}, \code{\link[RxnSim:ms.compute]{ms.compute}}
}